UI 3

+++ core/ui/src/commonMain/kotlin/ru/aleksandra/core/ui/model/Item.kt
@@ -2,8 +2,8 @@ package ru.aleksandra.core.ui.model
 
 data class Item(
     val name: String,
-    val priceWithoutDiscount: Float,
-    val priceWithDiscount: Float,
+    val priceWithoutDiscount: Int,
+    val priceWithDiscount: Int,
     val salePercent: Int,
     val count: Int,
     val image: String
